Changes made for files name and added the schema for action request
