+++ backend/actions/src/dto/actionsDto.dto.ts
@@ -1,10 +1,10 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { ActionBase } from "../models/actions_base";
+import { ActionBase } from "../models/actions_base.entity";
 
 export class ActionsDto {
     @ApiProperty({
         description: "Actions list",
-        example: [{ title: "Action Name", description: "Action description", isWeb: true, isMobile: true, isDesktop:true }],
+        example: [{ actionName: "Action Name", description: "Action description", isWeb: true, isMobile: true, isDesktop:true }], 
     })
-    actions: ActionBase[];
+    Actions: ActionBase;
 }
