--- backend/actions/src/models/actions_base.ts
@@ -1,7 +1,0 @@
-export interface ActionBase {
-    actionName: string;
-    description: string;
-    isWeb: boolean;
-    isMobile: boolean;
-    isDesktop: boolean;
-}
