+++ backend/actions/src/responses/actions.res.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { ActionBase } from "../models/actions_base";
-import {Actions_type} from "src/models/actions";
+import { ActionBase } from "../models/actions_base.entity";
+import {Action} from "src/models/actions.entity";
 
 export class ActionsResponse {
     @ApiProperty({ description: "List of Actions" })
@@ -8,6 +8,6 @@ export class ActionsResponse {
 }
 
 export class ActionResponse {
-    @ApiProperty({ description: "Accept schema" })
-    Action: Actions_type;
+    @ApiProperty({ description: "Action schema" })
+    Action: Action;
 }
