+++ backend/actions/test/app.e2e-spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { INestApplication } from "@nestjs/common";
 import * as request from "supertest";
-import { AppModule } from "./../src/app.module";
+import { ActionModule } from "../src/actions.module";
 
 describe("AppController (e2e)", () => {
     let app: INestApplication;
 
     beforeEach(async () => {
         const moduleFixture: TestingModule = await Test.createTestingModule({
-            imports: [AppModule],
+            imports: [ActionModule],
         }).compile();
 
         app = moduleFixture.createNestApplication();
